change key type from integer to double, rename db_services -> db

+++ backend/init_db.sql
@@ -1,14 +1,14 @@
 CREATE TABLE "user"(
-    chat_id INTEGER PRIMARY KEY,
+    chat_id DOUBLE PRIMARY KEY,
     username VARCHAR(255)
 );
 CREATE TABLE poll(
-    poll_id INTEGER PRIMARY KEY,
+    poll_id DOUBLE PRIMARY KEY,
     question VARCHAR(1023)
 );
 CREATE TABLE response(
-    chat_id INTEGER,
-    poll_id INTEGER,
+    chat_id DOUBLE,
+    poll_id DOUBLE,
     answer VARCHAR(255),
     FOREIGN KEY(chat_id) REFERENCES "user"(chat_id) ON DELETE CASCADE,
     FOREIGN KEY(poll_id) REFERENCES poll(poll_id) ON DELETE CASCADE
